RDKBDEV-655: syscfg: add additional syscfg_set() APIs
Add a range of new APIs to set string or integer values, with or
without a namespace and with or without a call to syscfg_commit().

There is a lot of duplicated code within RDKB to do common tasks such
as setting a syscfg value from an integer or setting a syscfg value
followed by calling syscfg_commit(). The RDKB code can be simplified
by providing APIs to perform these common tasks directly.

syscfg: update syscfg_set_u() to accept unsigned long instead of unsigned

A lot of RDKB code uses ULONG when it menas a 32bit unsigned value.
Update syscfg_set_u() so that it can accept the ULONG data type
without generating compiler warnings...

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I82c013a9f19a1d39a2bbdc5416d453442df4bee5